third_party/WebKit/LayoutTests/fast/dom/shadow/nodetype-expected.txt

This tests the shadow host's and shadow root's nodeType.

PASS progress.nodeType is Node.ELEMENT_NODE
PASS shadow.nodeType is Node.DOCUMENT_FRAGMENT_NODE
PASS shadowChild is non-null.
PASS shadowChild.nodeType != Node.DOCUMENT_FRAGMENT_NODE is true
PASS successfullyParsed is true

TEST COMPLETE
